+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Upload profile picture action from `uploadProfilePicture` and `updateProfilePicture` mutations.
 
 ## [2.113.5] - 2020-01-10
 ### Added